requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.6.2.post1
attrs==24.2.0
audioread==3.0.1
av==10.0.0
babel==2.16.0
boto3==1.35.54
botocore==1.35.54
cached-path==1.6.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chattts-fork==0.0.8
click==8.1.7
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
ctranslate2==3.24.0
cycler==0.12.1
cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
deprecated==1.2.14
distance==0.1.3
docopt==0.6.2
dtw-python==1.4.4
einops==0.8.0
einx==0.3.0
encodec==0.1.1
eng-to-ipa==0.0.2
fastapi==0.115.4
faster-whisper==0.9.0
ffmpy==0.4.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.6
fsspec==2024.2.0
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gradio==3.48.0
gradio-client==0.6.1
grpcio==1.67.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib-resources==6.4.5
inflect==7.0.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langid==1.1.6
librosa==0.9.1
llvmlite==0.43.0
loguru==0.7.2
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mecab-python3==1.0.9
melotts @ git+https://github.com/myshell-ai/MeloTTS.git@5b538481e24e0d578955be32a95d88fcbde26dc8
more-itertools==10.5.0
mpmath==1.3.0
myshell-openvoice @ git+https://github.com/myshell-ai/OpenVoice.git@f3cf835540572ade1460c8952f39d53e4f7952df
networkx==3.2.1
nltk==3.9.1
num2words==0.5.12
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.20.0
openai-whisper==20240930
orjson==3.10.11
packaging==24.1
pandas==1.5.1
pillow==10.4.0
pip==24.0
plac==1.4.3
platformdirs==4.3.6
pooch==1.8.2
proces==0.1.7
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydub==0.25.1
pygments==2.18.0
pykakasi==2.2.1
pyparsing==3.2.0
pypinyin==0.50.0
pyreadline3==3.5.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
pyyaml==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.41.2
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
toolz==1.0.0
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.6
transformers==4.41.2
txtsplit==1.0.0
typing-extensions==4.9.0
tzdata==2024.2
tzlocal==5.2
unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.2.3
uv==0.4.29
uvicorn==0.32.0
vector-quantize-pytorch==1.18.5
vocos==0.1.0
wasabi==0.10.1
wavmark==0.0.3
websockets==11.0.3
werkzeug==3.1.1
whisper-timestamped==1.14.2
win32-setctime==1.1.0
wrapt==1.16.0
